# Fuse adjacent raw data appends