--- a/gen/encode.go
+++ b/gen/encode.go
@@ -14,7 +14,8 @@ func encode(w io.Writer) *encodeGen {
 
 type encodeGen struct {
 	passes
-	p printer
+	p    printer
+	fuse []byte
 }
 
 func (e *encodeGen) Method() Method { return Encode }
@@ -26,6 +27,21 @@ func (e *encodeGen) Apply(dirs []string) error {
 func (e *encodeGen) writeAndCheck(typ string, argfmt string, arg interface{}) {
 	e.p.printf("\nerr = en.Write%s(%s)", typ, fmt.Sprintf(argfmt, arg))
 	e.p.print(errcheck)
+}
+
+func (e *encodeGen) fuseHook() {
+	if len(e.fuse) > 0 {
+		e.appendraw(e.fuse)
+		e.fuse = e.fuse[:0]
+	}
+}
+
+func (e *encodeGen) Fuse(b []byte) {
+	if len(e.fuse) > 0 {
+		e.fuse = append(e.fuse, b...)
+	} else {
+		e.fuse = b
+	}
 }
 
 func (e *encodeGen) Execute(p Elem) error {
@@ -64,7 +80,7 @@ func (e *encodeGen) tuple(s *Struct) {
 	nfields := len(s.Fields)
 	data := msgp.AppendArrayHeader(nil, uint32(nfields))
 	e.p.printf("\n// array header, size %d", nfields)
-	e.appendraw(data)
+	e.Fuse(data)
 	for i := range s.Fields {
 		if !e.p.ok() {
 			return
@@ -88,15 +104,14 @@ func (e *encodeGen) structmap(s *Struct) {
 	nfields := len(s.Fields)
 	data := msgp.AppendMapHeader(nil, uint32(nfields))
 	e.p.printf("\n// map header, size %d", nfields)
-	e.appendraw(data)
+	e.Fuse(data)
 	for i := range s.Fields {
 		if !e.p.ok() {
 			return
 		}
-		data = data[:0]
-		data = msgp.AppendString(data, s.Fields[i].FieldTag)
+		data = msgp.AppendString(nil, s.Fields[i].FieldTag)
 		e.p.printf("\n// write %q", s.Fields[i].FieldTag)
-		e.appendraw(data)
+		e.Fuse(data)
 		next(e, s.Fields[i].FieldElem)
 	}
 }
@@ -105,6 +120,7 @@ func (e *encodeGen) gMap(m *Map) {
 	if !e.p.ok() {
 		return
 	}
+	e.fuseHook()
 	vname := m.Varname()
 	e.writeAndCheck(mapHeader, lenAsUint32, vname)
 
@@ -118,7 +134,7 @@ func (e *encodeGen) gPtr(s *Ptr) {
 	if !e.p.ok() {
 		return
 	}
-
+	e.fuseHook()
 	e.p.printf("\nif %s == nil { err = en.WriteNil(); if err != nil { return; } } else {", s.Varname())
 	next(e, s.Value)
 	e.p.closeblock()
@@ -128,6 +144,7 @@ func (e *encodeGen) gSlice(s *Slice) {
 	if !e.p.ok() {
 		return
 	}
+	e.fuseHook()
 	e.writeAndCheck(arrayHeader, lenAsUint32, s.Varname())
 	e.p.rangeBlock(s.Index, s.Varname(), e, s.Els)
 }
@@ -136,6 +153,7 @@ func (e *encodeGen) gArray(a *Array) {
 	if !e.p.ok() {
 		return
 	}
+	e.fuseHook()
 	// shortcut for [const]byte
 	if be, ok := a.Els.(*BaseElem); ok && (be.Value == Byte || be.Value == Uint8) {
 		e.p.printf("\nerr = en.WriteBytes(%s[:])", a.Varname())
@@ -151,6 +169,7 @@ func (e *encodeGen) gBase(b *BaseElem) {
 	if !e.p.ok() {
 		return
 	}
+	e.fuseHook()
 	vname := b.Varname()
 	if b.Convert {
 		vname = tobaseConvert(b)