--- a/gen/marshal.go
+++ b/gen/marshal.go
@@ -14,7 +14,8 @@ func marshal(w io.Writer) *marshalGen {
 
 type marshalGen struct {
 	passes
-	p printer
+	p    printer
+	fuse []byte
 }
 
 func (m *marshalGen) Method() Method { return Marshal }
@@ -53,6 +54,21 @@ func (m *marshalGen) rawAppend(typ string, argfmt string, arg interface{}) {
 	m.p.printf("\no = msgp.Append%s(o, %s)", typ, fmt.Sprintf(argfmt, arg))
 }
 
+func (m *marshalGen) fuseHook() {
+	if len(m.fuse) > 0 {
+		m.rawbytes(m.fuse)
+		m.fuse = m.fuse[:0]
+	}
+}
+
+func (m *marshalGen) Fuse(b []byte) {
+	if len(m.fuse) == 0 {
+		m.fuse = b
+	} else {
+		m.fuse = append(m.fuse, b...)
+	}
+}
+
 func (m *marshalGen) gStruct(s *Struct) {
 	if !m.p.ok() {
 		return
@@ -70,7 +86,7 @@ func (m *marshalGen) tuple(s *Struct) {
 	data := make([]byte, 0, 5)
 	data = msgp.AppendArrayHeader(data, uint32(len(s.Fields)))
 	m.p.printf("\n// array header, size %d", len(s.Fields))
-	m.rawbytes(data)
+	m.Fuse(data)
 	for i := range s.Fields {
 		if !m.p.ok() {
 			return
@@ -83,16 +99,15 @@ func (m *marshalGen) mapstruct(s *Struct) {
 	data := make([]byte, 0, 64)
 	data = msgp.AppendMapHeader(data, uint32(len(s.Fields)))
 	m.p.printf("\n// map header, size %d", len(s.Fields))
-	m.rawbytes(data)
+	m.Fuse(data)
 	for i := range s.Fields {
 		if !m.p.ok() {
 			return
 		}
-		data = data[:0]
-		data = msgp.AppendString(data, s.Fields[i].FieldTag)
+		data = msgp.AppendString(nil, s.Fields[i].FieldTag)
 
 		m.p.printf("\n// string %q", s.Fields[i].FieldTag)
-		m.rawbytes(data)
+		m.Fuse(data)
 
 		next(m, s.Fields[i].FieldElem)
 	}
@@ -111,7 +126,7 @@ func (m *marshalGen) gMap(s *Map) {
 	if !m.p.ok() {
 		return
 	}
-
+	m.fuseHook()
 	vname := s.Varname()
 	m.rawAppend(mapHeader, lenAsUint32, vname)
 	m.p.printf("\nfor %s, %s := range %s {", s.Keyidx, s.Validx, vname)
@@ -124,6 +139,7 @@ func (m *marshalGen) gSlice(s *Slice) {
 	if !m.p.ok() {
 		return
 	}
+	m.fuseHook()
 	vname := s.Varname()
 	m.rawAppend(arrayHeader, lenAsUint32, vname)
 	m.p.rangeBlock(s.Index, vname, m, s.Els)
@@ -133,7 +149,7 @@ func (m *marshalGen) gArray(a *Array) {
 	if !m.p.ok() {
 		return
 	}
-
+	m.fuseHook()
 	if be, ok := a.Els.(*BaseElem); ok && be.Value == Byte {
 		m.rawAppend("Bytes", "%s[:]", a.Varname())
 		return
@@ -147,6 +163,7 @@ func (m *marshalGen) gPtr(p *Ptr) {
 	if !m.p.ok() {
 		return
 	}
+	m.fuseHook()
 	m.p.printf("\nif %s == nil {\no = msgp.AppendNil(o)\n} else {", p.Varname())
 	next(m, p.Value)
 	m.p.closeblock()
@@ -156,7 +173,7 @@ func (m *marshalGen) gBase(b *BaseElem) {
 	if !m.p.ok() {
 		return
 	}
-
+	m.fuseHook()
 	vname := b.Varname()
 
 	if b.Convert {