```
benchmark                         old ns/op     new ns/op     delta
BenchmarkFastEncode               83.6          80.7          -3.47%
BenchmarkFastDecode               127           123           -3.15%
BenchmarkThingsMarshalMsg         676           671           -0.74%
BenchmarkThingsAppendMsg          268           267           -0.37%
BenchmarkThingsUnmarshal          406           409           +0.74%
BenchmarkThingsEncode             320           307           -4.06%
BenchmarkThingsDecode             662           658           -0.60%
BenchmarkFilesMarshalMsg          76.3          76.4          +0.13%
BenchmarkFilesAppendMsg           17.1          17.2          +0.58%
BenchmarkFilesUnmarshal           13.5          13.4          -0.74%
BenchmarkFilesEncode              12.1          12.9          +6.61%
BenchmarkFilesDecode              20.7          21.0          +1.45%
BenchmarkTestFastMarshalMsg       238           243           +2.10%
BenchmarkTestFastAppendMsg        81.8          81.2          -0.73%
BenchmarkTestFastUnmarshal        77.1          77.0          -0.13%
BenchmarkTestFastEncode           79.1          79.1          +0.00%
BenchmarkTestFastDecode           114           113           -0.88%
BenchmarkCustomMarshalMsg         288           282           -2.08%
BenchmarkCustomAppendMsg          97.1          90.9          -6.39%
BenchmarkCustomUnmarshal          284           285           +0.35%
BenchmarkCustomEncode             155           124           -20.00%
BenchmarkCustomDecode             357           356           -0.28%
BenchmarkTestTypeMarshalMsg       479           474           -1.04%
BenchmarkTestTypeAppendMsg        175           169           -3.43%
BenchmarkTestTypeUnmarshal        568           566           -0.35%
BenchmarkTestTypeEncode           259           228           -11.97%
BenchmarkTestTypeDecode           732           728           -0.55%
BenchmarkFileHandleMarshalMsg     163           161           -1.23%
BenchmarkFileHandleAppendMsg      42.8          40.9          -4.44%
BenchmarkFileHandleUnmarshal      110           109           -0.91%
BenchmarkFileHandleEncode         62.7          49.8          -20.57%
BenchmarkFileHandleDecode         125           120           -4.00%
BenchmarkTestHiddenMarshalMsg     126           124           -1.59%
BenchmarkTestHiddenAppendMsg      37.8          35.7          -5.56%
BenchmarkTestHiddenUnmarshal      112           112           +0.00%
BenchmarkTestHiddenEncode         59.0          49.2          -16.61%
BenchmarkTestHiddenDecode         123           122           -0.81%
BenchmarkEmbeddedMarshalMsg       256           253           -1.17%
BenchmarkEmbeddedAppendMsg        67.0          64.8          -3.28%
BenchmarkEmbeddedUnmarshal        207           205           -0.97%
BenchmarkEmbeddedEncode           95.7          84.9          -11.29%
BenchmarkEmbeddedDecode           260           255           -1.92%
BenchmarkBlockMarshalMsg          243           243           +0.00%
BenchmarkBlockAppendMsg           29.5          29.3          -0.68%
BenchmarkBlockUnmarshal           17.7          17.5          -1.13%
BenchmarkBlockEncode              22.6          22.7          +0.44%
BenchmarkBlockDecode              32.0          32.2          +0.63%
BenchmarkInsaneMarshalMsg         193           189           -2.07%
BenchmarkInsaneAppendMsg          61.0          61.0          +0.00%
BenchmarkInsaneUnmarshal          40.5          40.9          +0.99%
BenchmarkXMarshalMsg              498           492           -1.20%
BenchmarkXAppendMsg               80.5          78.4          -2.61%
BenchmarkXUnmarshal               184           185           +0.54%
BenchmarkXEncode                  113           103           -8.85%
BenchmarkXDecode                  279           279           +0.00%
BenchmarkTestBenchMarshalMsg      266           266           +0.00%
BenchmarkTestBenchAppendMsg       98.9          98.7          -0.20%
BenchmarkTestBenchUnmarshal       142           142           +0.00%
BenchmarkTestBenchEncode          103           103           +0.00%
BenchmarkTestBenchDecode          148           150           +1.35%

benchmark                        old MB/s     new MB/s     speedup
BenchmarkFastEncode              526.29       545.06       1.04x
BenchmarkFastDecode              344.12       356.01       1.03x
BenchmarkThingsAppendMsg         525.90       527.24       1.00x
BenchmarkThingsUnmarshal         346.45       343.92       0.99x
BenchmarkThingsEncode            439.72       458.79       1.04x
BenchmarkThingsDecode            212.77       214.15       1.01x
BenchmarkFilesAppendMsg          58.34        58.30        1.00x
BenchmarkFilesUnmarshal          74.05        74.46        1.01x
BenchmarkFilesEncode             82.82        77.24        0.93x
BenchmarkFilesDecode             48.38        47.58        0.98x
BenchmarkTestFastAppendMsg       366.72       369.58       1.01x
BenchmarkTestFastUnmarshal       388.95       389.76       1.00x
BenchmarkTestFastEncode          379.41       379.46       1.00x
BenchmarkTestFastDecode          261.76       264.56       1.01x
BenchmarkCustomAppendMsg         381.14       407.17       1.07x
BenchmarkCustomUnmarshal         129.97       129.52       1.00x
BenchmarkCustomEncode            238.41       298.31       1.25x
BenchmarkCustomDecode            103.43       103.81       1.00x
BenchmarkTestTypeAppendMsg       523.62       542.29       1.04x
BenchmarkTestTypeUnmarshal       161.74       162.44       1.00x
BenchmarkTestTypeEncode          354.39       402.99       1.14x
BenchmarkTestTypeDecode          125.54       126.30       1.01x
BenchmarkFileHandleAppendMsg     327.42       342.54       1.05x
BenchmarkFileHandleUnmarshal     126.23       128.40       1.02x
BenchmarkFileHandleEncode        223.20       280.87       1.26x
BenchmarkFileHandleDecode        111.15       116.19       1.05x
BenchmarkTestHiddenAppendMsg     185.08       195.81       1.06x
BenchmarkTestHiddenUnmarshal     62.10        62.37        1.00x
BenchmarkTestHiddenEncode        118.59       142.30       1.20x
BenchmarkTestHiddenDecode        56.83        57.13        1.01x
BenchmarkEmbeddedAppendMsg       612.12       632.47       1.03x
BenchmarkEmbeddedUnmarshal       197.99       199.68       1.01x
BenchmarkEmbeddedEncode          428.25       483.02       1.13x
BenchmarkEmbeddedDecode          157.14       160.47       1.02x
BenchmarkBlockAppendMsg          1152.02      1158.56      1.01x
BenchmarkBlockUnmarshal          1918.00      1942.12      1.01x
BenchmarkBlockEncode             1502.55      1495.12      1.00x
BenchmarkBlockDecode             1063.74      1056.76      0.99x
BenchmarkInsaneAppendMsg         65.62        65.54        1.00x
BenchmarkInsaneUnmarshal         98.88        97.68        0.99x
BenchmarkXAppendMsg              1205.40      1236.53      1.03x
BenchmarkXUnmarshal              524.63       524.30       1.00x
BenchmarkXEncode                 852.95       940.89       1.10x
BenchmarkXDecode                 346.52       347.20       1.00x
BenchmarkTestBenchAppendMsg      293.15       293.74       1.00x
BenchmarkTestBenchUnmarshal      203.74       203.69       1.00x
BenchmarkTestBenchEncode         280.97       279.81       1.00x
BenchmarkTestBenchDecode         195.13       192.92       0.99x
```

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/tinylib/msgp/97)
<!-- Reviewable:end -->